setup interface and abstract class framework
