+++ docs/com/github/zipcodewilmington/casino/games/slots/SlotsGame.html
@@ -93,7 +93,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.github.zipcodewilmington.casino.games.slots.SlotsGame</li>
+<li>com.github.zipcodewilmington.casino.games.gamblingGames.slots.SlotsGame</li>
 </ul>
 </li>
 </ul>
