+++ docs/com/github/zipcodewilmington/casino/games/slots/SlotsPlayer.html
@@ -93,7 +93,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.github.zipcodewilmington.casino.games.slots.SlotsPlayer</li>
+<li>com.github.zipcodewilmington.casino.games.gamblingGames.slots.SlotsPlayer</li>
 </ul>
 </li>
 </ul>
