+++ src/main/java/com/github/zipcodewilmington/casino/GameInterface.java
@@ -27,4 +27,11 @@ public interface GameInterface extends Runnable {
      * specifies how the game will run
      */
     void run();
+
+     void intro();
+
+     boolean winCheck();
+
+     void quitAsk();
+
 }
