--- src/main/java/com/github/zipcodewilmington/casino/games/BlackJack/BlackJackGame.java
@@ -1,4 +1,0 @@
-package com.github.zipcodewilmington.casino.games.BlackJack;
-
-public class BlackJackGame {
-}
