--- src/main/java/com/github/zipcodewilmington/casino/games/Craps/CrapsPlayer.java
@@ -1,4 +1,0 @@
-package com.github.zipcodewilmington.casino.games.Craps;
-
-public class CrapsPlayer {
-}
