--- src/main/java/com/github/zipcodewilmington/casino/games/RockPaperScissors/RockPaperScissorsGame.java
@@ -1,4 +1,0 @@
-package com.github.zipcodewilmington.casino.games.RockPaperScissors;
-
-public class RockPaperScissorsGame {
-}
