--- src/main/java/com/github/zipcodewilmington/casino/games/Roulette/RouletteGame.java
@@ -1,4 +1,0 @@
-package com.github.zipcodewilmington.casino.games.Roulette;
-
-public class RouletteGame {
-}
