--- src/main/java/com/github/zipcodewilmington/casino/games/Roulette/RoulettePlayer.java
@@ -1,4 +1,0 @@
-package com.github.zipcodewilmington.casino.games.Roulette;
-
-public class RoulettePlayer {
-}
