--- src/main/java/com/github/zipcodewilmington/casino/games/WordGuess/WordGuessPlayer.java
@@ -1,4 +1,0 @@
-package com.github.zipcodewilmington.casino.games.WordGuess;
-
-public class WordGuessPlayer {
-}
