+++ src/main/java/com/github/zipcodewilmington/casino/games/casualGames/wordGuess/CasualGame.java
@@ -1,0 +1,6 @@
+package com.github.zipcodewilmington.casino.games.casualGames.wordGuess;
+
+import com.github.zipcodewilmington.utils.Game;
+
+public abstract class CasualGame extends Game {
+}
