+++ src/main/java/com/github/zipcodewilmington/casino/games/casualGames/wordGuess/WordGuessGame.java
@@ -1,0 +1,4 @@
+package com.github.zipcodewilmington.casino.games.casualGames.wordGuess;
+
+public class WordGuessGame {
+}
