+++ src/main/java/com/github/zipcodewilmington/casino/games/gamblingGames/BlackJack/BlackJackPlayer.java
@@ -1,0 +1,4 @@
+package com.github.zipcodewilmington.casino.games.gamblingGames.BlackJack;
+
+public class BlackJackPlayer {
+}
