+++ src/main/java/com/github/zipcodewilmington/casino/games/gamblingGames/Craps/CrapsPlayer.java
@@ -1,0 +1,4 @@
+package com.github.zipcodewilmington.casino.games.gamblingGames.Craps;
+
+public class CrapsPlayer {
+}
