+++ src/main/java/com/github/zipcodewilmington/casino/games/gamblingGames/GamblingGame.java
@@ -1,0 +1,6 @@
+package com.github.zipcodewilmington.casino.games.gamblingGames;
+
+import com.github.zipcodewilmington.utils.Game;
+
+public abstract class GamblingGame extends Game implements GamblingGameInterface{
+}
