+++ src/main/java/com/github/zipcodewilmington/casino/games/gamblingGames/GamblingGameInterface.java
@@ -1,0 +1,11 @@
+package com.github.zipcodewilmington.casino.games.gamblingGames;
+
+public interface GamblingGameInterface {
+
+    int playerBet();
+
+    int calculateWinner();
+
+    void updateBalance();
+
+}
