+++ src/main/java/com/github/zipcodewilmington/casino/games/gamblingGames/RockPaperScissors/RockPaperScissorsGame.java
@@ -1,0 +1,4 @@
+package com.github.zipcodewilmington.casino.games.gamblingGames.RockPaperScissors;
+
+public class RockPaperScissorsGame {
+}
