+++ src/main/java/com/github/zipcodewilmington/casino/games/gamblingGames/Roulette/RoulettePlayer.java
@@ -1,0 +1,4 @@
+package com.github.zipcodewilmington.casino.games.gamblingGames.Roulette;
+
+public class RoulettePlayer {
+}
