--- src/main/java/com/github/zipcodewilmington/casino/games/numberguess/NumberGuessPlayer.java
@@ -1,7 +1,0 @@
-package com.github.zipcodewilmington.casino.games.numberguess;
-
-/**
- * Created by leon on 7/21/2020.
- */
-public class NumberGuessPlayer {
-}
